Update README.md
A histogram is like a bar graph that visually represents how many pixels in a picture correspond to each shade of brightness, from the darkest to the lightest.  It helps us understand if an image is predominantly dark, bright, or has a good balance of both.  When a picture appears dull due to low contrast, histogram equalization comes to the rescue.  This process stretches out the brightness levels, effectively spreading the shades of gray more evenly.  The result is a more vibrant image with enhanced details, particularly in areas that were previously too dark or too bright.  In essence, histogram equalization optimizes the brightness range of an image, making it clearer and more visually appealing.